requirements.txt
spotipy==2.22.1
google-api-python-client==2.108.0
pymongo==4.6.0
python-dotenv==1.0.0
pandas==2.1.4
streamlit==1.28.2
plotly==5.17.0